+++ backend/migrations/000010_create_user_profiles.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS user_profiles (
-    user_id INTEGER PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
     favorite_genre_ids JSONB DEFAULT '[]',
     favorite_artist_ids JSONB DEFAULT '[]',
     starter_track_ids JSONB DEFAULT '[]',
