+++ backend/migrations/000011_create_events.up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS events (
     id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     entity_type VARCHAR NOT NULL,
-    entity_id INTEGER NOT NULL,
+    entity_id BIGINT NOT NULL,
     event_type VARCHAR NOT NULL,
     session_id VARCHAR,
     metadata JSONB,
