+++ backend/migrations/000012_create_recommendation_requests.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS recommendation_requests (
     id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     request_type VARCHAR NOT NULL,
     created_at TIMESTAMPTZ DEFAULT now()
 );
